move granular functions to trait
